refactor: Notifications in prebuilt (#3307)
